requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
faiss-cpu
sentence-transformers
requests
